PowerHAL: fix native crash when switching different profiles in a binder call
Native crash is caused by AdpfProfile was switching to another during a
binder call. The first one is HBoostEnabled and the 2nd is not. That
leads the code hit the Null Pointer Exception.

To fix the problem, we use a lock to make sure the setAdpfProfile and binder
calls running mutual exclusively.

To achieve the purpose,
1. HintManager supports register AdpfUpdateEvent.
2. PowerHintSession, instead of directly accessing the
   HintManager::GetAdpfProfile, it keeps its own mAdpfProfile and use a
   lock to prevent the profile from being changed during a binder call
   session.

Bug: 364903396
Test: manually check running YT all night long without libperf
      restarted.
Change-Id: I5e9b67b855ccaa0096e2b6639bca1b32b5db27e5