ADPF: Don't boost after the heuristic boost state gets reset
Making a large change to the target duration will cause the heuristic
boost session to get reset. Some users, particularly SurfaceFlinger, do
this frequently. Don't boost to the highest heuristic boost level for
sessions that have been reset in this way. The boost should only occur
when the session is initially created.

Bug: 395187033
Test: atest libadpf_test
Test: Verify with Perfetto traces that SurfaceFlinger isn't getting
boosted all the time

Change-Id: Iea0895d4b4fc93edab02799c6dbd0d9556564a66